[Fix] Rotation visual glitch.

+++ Game/Source/Archetypes/PlayerArchetype.cpp
@@ -1,5 +1,8 @@
 ﻿#include "pch.h"
 #include "Archetypes/PlayerArchetype.h"
+
+#include <iostream>
+
 #include "Systems/MovementSystem.h"
 #include "Systems/ResourceManager.h"
 #include "Systems/TurnSystem.h"
@@ -53,7 +56,7 @@ namespace game
 
 				// Round the from position.
 				const glm::vec2 from = glm::vec2(glm::ivec2(entity.transform.position));
-				const glm::vec2 delta = glm::vec2(dir) * static_cast<float>(vke::PIXEL_SIZE_ENTITY);
+				const glm::vec2 delta = glm::vec2(dir * static_cast<int32_t>(vke::PIXEL_SIZE_ENTITY));
 
 				movementComponent.from = from;
 				movementComponent.to = from + delta;
